feat(k8s-docker): polaris scanner dockerfile and k8s objects
DISCLAIMER:
I haven't found a way to run the service in bare-bone alpine container.

polaris keeps a part of runtime needed configuration in yaml files.
These files are embedded into polaris binaries with gobuffalo/packr
The problem is, packr is not able to embed files from referenced modules
https://github.com/gobuffalo/packr/pull/236

To overcome the issue, polaris-scanner is built and is running from the
same golang container golang:1.13.7-alpine3.11 using CGO_ENABLED=1
If produced binary is copied to bare-bone alpine container -
it's not visible, due the problem described in SO question
https://stackoverflow.com/q/34729748/1952990

I tried both options from SO question: nor symlink, neither
CGO_ENABLED=0 helps, because of the initial packr issue